Refactor: add KeepAwakeModifier and apply to priming/activation to keep screen awake when visible or busy. Reusable on other screens too if needed placed in Views/UIHelpers.swift
