+++ MedtrumKitUI/Views/Onboarding/PatchPrimingView.swift
@@ -100,30 +100,7 @@ struct PatchPrimingView: View {
         .edgesIgnoringSafeArea(.bottom)
         .navigationBarBackButtonHidden(viewModel.isPriming)
         .navigationTitle(LocalizedString("Patch priming", comment: "Priming header"))
-
-        // Visibility lifecycle
-        .onAppear {
-            isVisible = true
-            applyIdleTimerPolicy()
-        }
-        .onDisappear {
-            isVisible = false
-            applyIdleTimerPolicy()
-        }
-        .onChange(of: viewModel.isPriming) { _ in
-            applyIdleTimerPolicy()
-        }
-    }
-
-    private func applyIdleTimerPolicy() {
-        let shouldKeepAwake = isVisible || viewModel.isPriming
-        if Thread.isMainThread {
-            UIApplication.shared.isIdleTimerDisabled = shouldKeepAwake
-        } else {
-            DispatchQueue.main.async {
-                UIApplication.shared.isIdleTimerDisabled = shouldKeepAwake
-            }
-        }
+        .keepScreenAwake(whenBusy: viewModel.isPriming, alsoWhenVisible: true)
     }
 
     @ViewBuilder func supportImage(_ imageName: String) -> some View {
